Added gitignores to the repository.
Each of the yaffs direct tests has its own gitignore file.
This is so that if the test is moved or deleted then the
gitignore file is dealt with appropriately.

However all the symlinks are handled by the root .gitinore file.
If a test is moved or deleated, run the script in the root gitignore
file to update the list of all the symlinks in the repository.

Signed-off-by: Timothy Manning <codedraftsman@gmail.com>